+++ packages/core/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @atlaskit/pragmatic-drag-and-drop
 
+## 1.7.2
+
+### Patch Changes
+
+- [#164244](https://bitbucket.org/atlassian/atlassian-frontend-monorepo/pull-requests/164244)
+  [`65021fc0267e2`](https://bitbucket.org/atlassian/atlassian-frontend-monorepo/commits/65021fc0267e2) -
+  The cleanup functions returned by the following utilities now only work on the first call. This
+  was done to prevent unexpected side effects of calling a cleanup function multiple times.
+
+  - `@atlaskit/pragmatic-drag-and-drop/adapter/element`
+    - `draggable`
+    - `dropTargetForElements`
+    - `monitorForElements`
+  - `@atlaskit/pragmatic-drag-and-drop/adapter/text-selection`
+    - `dropTargetForTextSelection`
+    - `monitorForTextSelection`
+  - `@atlaskit/pragmatic-drag-and-drop/adapter/external`
+    - `dropTargetForExternal`
+    - `monitorForExternal`
+  - `@atlaskit/pragmatic-drag-and-drop-auto-scroll/element`
+    - `autoScrollForElements`
+    - `autoScrollWindowForElements`
+  - `@atlaskit/pragmatic-drag-and-drop-auto-scroll/external`
+    - `autoScrollForExternal`
+    - `autoScrollWindowForExternal`
+  - `@atlaskit/pragmatic-drag-and-drop-auto-scroll/text-selection`
+    - `autoScrollForTextSelection`
+    - `autoScrollWindowForTextSelection`
+
 ## 1.7.1
 
 ### Patch Changes